requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
pytesseract
rapidfuzz
opencv-python-headless